requirements.txt
backoff==2.2.1
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
Flask==2.3.1
googleapis-common-protos==1.56.2
grpcio==1.54.0
idna==3.4
importlib-metadata==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
opentelemetry-api==1.17.0
opentelemetry-distro==0.38b0
opentelemetry-exporter-jaeger==1.17.0
opentelemetry-exporter-jaeger-proto-grpc==1.17.0
opentelemetry-exporter-jaeger-thrift==1.17.0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-exporter-otlp-proto-grpc==1.17.0
opentelemetry-exporter-otlp-proto-http==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-aws-lambda==0.38b0
opentelemetry-instrumentation-dbapi==0.38b0
opentelemetry-instrumentation-flask==0.38b0
opentelemetry-instrumentation-jinja2==0.38b0
opentelemetry-instrumentation-logging==0.38b0
opentelemetry-instrumentation-sqlite3==0.38b0
opentelemetry-instrumentation-urllib==0.38b0
opentelemetry-instrumentation-wsgi==0.38b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
protobuf==3.20.3
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg-pool==3.1.7
requests==2.29.0
six==1.16.0
thrift==0.16.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.0
wrapt==1.15.0
zipp==3.15.0
